+++ bdew_datetimes/__init__.py
@@ -1,0 +1,6 @@
+"""
+bdew_datetimes is a package that models the BDEW holiday, which is relevant for German utilities
+"""
+from .calendar import create_bdew_calendar
+
+__version__ = "0.1"
