+++ tox.ini
@@ -26,7 +26,7 @@ deps =
     -r requirements.txt
     pylint
 commands =
-    pylint energy_datetime
+    pylint bdew_datetimes
 
 [testenv:type_check]
 # the type_check environment checks the type hints using mypy
@@ -34,7 +34,7 @@ deps =
     -r requirements.txt
     mypy
 commands =
-    mypy --show-error-codes energy_datetime
+    mypy --show-error-codes bdew_datetimes
     mypy --show-error-codes tests
 
 [testenv:format]
